Add derived datatype features
Derived datatypes are useful to communicate data with one or more structures/classes.
To use it, new files derive_datatype.hpp, access.hpp, and datatype_committer.hpp are created.
Moreover I introduce a wrapper for MPI_Datatype, and modify metafunctions to check if datatype exists and to get datatype.
Due to this modifications, some other files are modified.
There are files with old versions of metafunctions, but they will be removed soon.